+++ ReleaseNotes.md
@@ -7,6 +7,9 @@
 
 * Add Show*THING*Range configuration options to allow disabling
   highlighting of *THING* completely.
+* Integration with the Prismatic Tools mod so that the range of
+  the prismatic sprinkler matches the configured value in
+  Prismatic Tools.
 
 ### API Changes